- Suppression de la signature sur une demande de transferts accueil lorsque la décision n'est pas définitive (controller et xsl) - Correction d'un BUG losque l'on est en mode debug sur le transtipage String-->Interger (Integer.parseInt(extract_bea) --> Long.parseLong(extract_bea))
git-svn-id: https://subversion.renater.fr/anonscm/svn/esup-transferts/trunk/esup-transferts@238 3ddcba78-ea86-4027-a5a7-99004716dbae

+++ esup-transferts-utils/src/main/java/org/esupportail/transferts/utils/CheckBEA23.java
@@ -45,7 +45,8 @@ public class CheckBEA23 {
 			logger.debug("############### numero.length() ############ --> "+ numero.length());
 			logger.debug("############### lettreCle ############ --> "+ lettreCle);
 			logger.debug("############### extract_bea ############ --> "+ extract_bea);
-			logger.debug("############### Integer.parseInt(extract_bea) ############ --> "+ Integer.parseInt(extract_bea));
+//			logger.debug("############### Integer.parseInt(extract_bea) ############ --> "+ Integer.parseInt(extract_bea));
+			logger.debug("############### Integer.parseInt(extract_bea) ############ --> "+ Long.parseLong(extract_bea));
 			logger.debug("############### modulo_bea ############ --> "+ modulo_bea);
 			logger.debug("############### cleCalc ############ --> "+ cleCalc);
 			logger.debug("############### lettreCle ############ --> "+ lettreCle);
